requirements.txt
streamlit
openai-whisper
torch
numpy
librosa
sounddevice